+++ RiskOfChaos/EffectDefinitions/World/HoldoutZone/IncreaseHoldoutZoneRadius.cs
@@ -1,0 +1,42 @@
+﻿using RiskOfChaos.ConfigHandling;
+using RiskOfChaos.EffectHandling;
+using RiskOfChaos.EffectHandling.EffectClassAttributes;
+using RiskOfChaos.EffectHandling.EffectClassAttributes.Data;
+using RiskOfChaos.EffectHandling.EffectClassAttributes.Methods;
+using RiskOfOptions.OptionConfigs;
+
+namespace RiskOfChaos.EffectDefinitions.World.HoldoutZone
+{
+    [ChaosTimedEffect("increase_holdout_zone_radius", TimedEffectType.UntilStageEnd, ConfigName = "Increase Teleporter Zone Radius")]
+    public sealed class IncreaseHoldoutZoneRadius : GenericHoldoutZoneModifierEffect
+    {
+        [EffectConfig]
+        static readonly ConfigHolder<float> _radiusIncrease =
+            ConfigFactory<float>.CreateConfig("Radius Increase", 0.5f)
+                                .Description("Percentage increase of teleporter radius")
+                                .OptionConfig(new StepSliderConfig
+                                {
+                                    formatString = "+{0:P0}",
+                                    increment = 0.01f,
+                                    min = 0f,
+                                    max = 2f
+                                })
+                                .ValueConstrictor(CommonValueConstrictors.GreaterThanOrEqualTo(0f))
+                                .Build();
+
+        [EffectNameFormatArgs]
+        static object[] GetEffectNameFormatArgs()
+        {
+            return new object[]
+            {
+                _radiusIncrease.Value
+            };
+        }
+
+        protected override void modifyRadius(ref float radius)
+        {
+            base.modifyRadius(ref radius);
+            radius *= 1f + _radiusIncrease.Value;
+        }
+    }
+}